+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- Add new metric type: object ([#587](https://github.com/mozilla/glean_parser/pull/587))
+- Add new metric type object (only Rust codegen support right now) ([#587](https://github.com/mozilla/glean_parser/pull/587))
 
 ## 11.1.0